Finish rudimetary HTTP server
AsyncServer is an asynchronous server.
AsyncClient is an asynchronous client connection.
Both need a lot of work to be good, but they work for now.

HttpServer is an asynchronous server that will provide classical
webserver-functionality. As of now, it's barely more than a proxy-class
for AsyncServer.

HttpRequest and HttpResponse are the basic building blocks for HTTP-
communication.