Configure cmake as the primary build system
SUC is built once as an executable (main.cpp), which I'll probably
remove since test_client and test_server will act as my testing
programs, and as a static library.

test/ contains test programs for a server and a client. The tests use
the static library built in SocketUtilityC++/CMakeLists.txt.